postgres/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# postgres/client.py: 4
PyYAML == 5.1.2

# postgres/client.py: 5
psycopg2 == 2.8.5
